Version I am leaving with right now.  It seems to work ok.  Leery about the height controller.  Also when I implement the new safety function the robot tends to reset when running backwards along a trajectory.  I believe it resets because the robot pose errors get too large.  Behavior is consistent with that too, it seems the robot has to go much faster on a negative path.  Very weird.

+++ prototypes_al_pupp.h
@@ -65,7 +65,7 @@ void SetPose(float xdest, float ydest, float thdest);
 void RuntimeOperation(void);    			
 void reset_robot(void);
 // Functions for kinematic controller:
-void run_filo(const double new_val, double *array);
+void run_filo(const float new_val, float *array);
 void calculate_controller_gains(void);
 int calculate_feedforward_values(const float k);
 void setup_controller(void);
